backend/requirements.txt
flask==2.2.3
flask-cors==3.0.10
requests==2.28.2
beautifulsoup4==4.12.0
python-dotenv==1.0.0
werkzeug==2.2.3 